Modify assignment3
