--- week4/assignment3/src/componments/Hidedcontainer.jsx
@@ -1,20 +1,0 @@
-import React from "react";
-
-export default function Hidedcontainer({ action }) {
-    return (
-        <div className={`box-container-hided ${action ? "hide" : "show"}`}>
-            <div id='content-box' className='content-box'>
-                <p className='box'>Content Box 5</p>
-            </div>
-            <div id='content-box' className='content-box'>
-                <p className='box'>Content Box 6</p>
-            </div>
-            <div id='content-box' className='content-box'>
-                <p className='box'>Content Box 7</p>
-            </div>
-            <div id='content-box' className='content-box'>
-                <p className='box'>Content Box 8</p>
-            </div>
-        </div>
-    );
-}
